refact user interaction
