+++ sandbox/tile/gui/move.py
@@ -1,12 +1,15 @@
 # coding: utf-8
+import typing
+
+from synergine2_cocos2d.interaction import BaseActorInteraction
 from sandbox.tile.user_action import UserAction
-from synergine2.terminals import TerminalPackage
+from synergine2.simulation import SimulationBehaviour
+from synergine2_cocos2d.actor import Actor
 from synergine2_cocos2d.gl import draw_line
-from synergine2_cocos2d.interaction import Interaction
 from synergine2_xyz.move.simulation import RequestMoveBehaviour
 
 
-class BaseMoveActorInteraction(Interaction):
+class BaseMoveActorInteraction(BaseActorInteraction):
     gui_action = None
     color = None
     request_move_behaviour_class = RequestMoveBehaviour
@@ -22,27 +25,12 @@ class BaseMoveActorInteraction(Interaction):
                 self.color,
             )
 
-    def get_package_for_terminal(self) -> TerminalPackage:
-        # TODO: FinishMoveEvent ?
-        actions = []
-        mouse_grid_position = self.layer_manager.grid_manager.get_grid_position(
-            self.layer_manager.scrolling_manager.screen_to_world(
-                *self.layer_manager.edit_layer.screen_mouse,
-            )
-        )
-
-        for actor in self.layer_manager.edit_layer.selection:
-            actions.append((
-                self.request_move_behaviour_class, {
-                    'subject_id': actor.subject.id,
-                    'move_to': mouse_grid_position,
-                    'gui_action': self.gui_action,
-                }
-            ))
-
-        return TerminalPackage(
-            simulation_actions=actions,
-        )
+    def get_behaviour(self, actor: Actor, mouse_grid_position) -> typing.Tuple[typing.Type[SimulationBehaviour], dict]:
+        return self.request_move_behaviour_class, {
+            'subject_id': actor.subject.id,
+            'move_to': mouse_grid_position,
+            'gui_action': self.gui_action,
+        }
 
 
 class MoveActorInteraction(BaseMoveActorInteraction):
